% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closesamples.R
\name{GetClosestSamples}
\alias{GetClosestSamples}
\title{Get n samples}
\usage{
GetClosestSamples(
  n,
  phy_full,
  taxa_feasible,
  replace_full = TRUE,
  replace_feasible = FALSE,
  truncate_full_to_mrca = FALSE,
  less_memory = FALSE,
  descendant_labeled = FALSE,
  fast_ultrametric = FALSE
)
}
\arguments{
\item{n}{How many taxa to sample}

\item{phy_full}{A phylo object with all feasible taxa to sample from}

\item{taxa_feasible}{A vector of taxon names that are feasible to study (often on another tree)}

\item{replace_full}{If TRUE, will allow selecting the same taxon from the full tree more than once. If FALSE, forbids this. Both cases return n taxa}

\item{replace_feasible}{If TRUE, will allow selecting the same taxon from the set of feasible taxa more than once (returning a smaller than desired tree). If FALSE, forbids this.}

\item{truncate_full_to_mrca}{If TRUE, prune the full tree to the node that is the MRCA of the}

\item{less_memory}{If TRUE, uses a much slower approach that will not create giant matrices}

\item{descendant_labeled}{If TRUE, assumes the phy_full has been labeled with LabelNodesWithFeasibleDescendants}

\item{fast_ultrametric}{If TRUE, uses a fast algorithm for ultrametric trees}
}
\value{
A data.frame of chosen taxa, closest feasible match, and distance between them
}
\description{
Get n samples
}
