% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closesamples.R
\name{GetEnclosingTaxonomy}
\alias{GetEnclosingTaxonomy}
\title{Get enclosing taxonomy for a focal tree}
\usage{
GetEnclosingTaxonomy(phy_focal)
}
\arguments{
\item{tips}{A vetor of taxon names}
}
\value{
A phylogeny of the enclosing taxonomy
}
\description{
If you don't have a taxonomy tree for the focal tree, this will find one
}
