% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closesamples.R
\name{LabelNodesWithChosenDescendants}
\alias{LabelNodesWithChosenDescendants}
\title{Include descendant taxon ids in node labels}
\usage{
LabelNodesWithChosenDescendants(node, phy, clean = FALSE, sep = ", ")
}
\arguments{
\item{node}{node number, typically a tip}

\item{phy}{phylo object}

\item{clean}{wipe existing node labels}

\item{sep}{Separator}
}
\value{
A phylogeny with terminal node numbers as node.labels. Unfortunately, each will start with NA.
}
\description{
Include descendant taxon ids in node labels
}
