% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closesamples.R
\name{LabelNodesWithFeasibleDescendants}
\alias{LabelNodesWithFeasibleDescendants}
\title{Include descendant taxon ids in node labels}
\usage{
LabelNodesWithFeasibleDescendants(taxa_feasible, phy)
}
\arguments{
\item{taxa_feasible}{vector of taxon names that are feasible for later suse}

\item{phy}{phylo object}
}
\value{
A phylogeny with terminal node numbers as node.labels.
}
\description{
Include descendant taxon ids in node labels
}
