% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closesamples.R
\name{GetClosestSamples}
\alias{GetClosestSamples}
\title{Get n samples}
\usage{
GetClosestSamples(n, phy_full, taxa_possible, replace = TRUE)
}
\arguments{
\item{n}{How many taxa to sample}

\item{phy_full}{A phylo object with all possible taxa to sammple from}

\item{taxa_possible}{A vector of taxon names that are possible to study (often on another tree)}
}
\value{
A data.frame of chosen taxa, closest possible match, and distance between them
}
\description{
Get n samples
}
