% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closesamples.R
\name{SubsampleTree}
\alias{SubsampleTree}
\title{Return a sampled tree}
\usage{
SubsampleTree(phy_focal, n, phy_full = NULL, replace = TRUE)
}
\arguments{
\item{phy_focal}{The tree to subsample}

\item{n}{The number of taxa to include}

\item{phy_full}{The larger tree giving relationships}
}
\value{
A phylo object where taxa are sampled based on representing flat sampling from taxonomy
}
\description{
Return a sampled tree
}
